pyproject.toml
[project]
name = "outpatient-efficiency-dashboard"
version = "0.2.0"
description = "外来効率化ダッシュボード（東京医療センター経営企画室）"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.3",
    "black>=24.0",
]

[project.scripts]
outpatient-dashboard = "src.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
